patch: image-chat
