+++ src/components/inputBar.tsx
@@ -6,6 +6,7 @@ import {
   Dispatch,
   FormEvent,
   SetStateAction,
+  useCallback,
   useEffect,
   useState,
 } from "react";
@@ -24,6 +25,7 @@ import { useImageState } from "@/store/tlDrawImage";
 // import ModelSwitcher from "./modelswitcher";
 // import VadAudio from "./vadAudio";
 import VadAudio from "./VadAudio";
+import { useQueryState } from "next-usequerystate";
 const isValidImageType = (value: string) =>
   /^image\/(jpeg|png|jpg|webp)$/.test(value);
 
@@ -100,7 +102,163 @@ const InputBar = (props: InputBarProps) => {
   const [disableInputs, setDisableInputs] = useState<boolean>(false);
   const [isRagLoading, setIsRagLoading] = useState<boolean>(false);
   const queryClient = useQueryClient();
+  const [isNewChat, setIsNewChat] = useQueryState("new");
+  const [isFromClipboard, setIsFromClipboard] = useQueryState("clipboard");
+  const [incomingModel] = useQueryState("model");
+  const [incomingInput] = useQueryState("input");
+  const [chattype, setChattype] = useState<ChatType>(
+    props?.chattype || incomingModel || "chat",
+  );
 
+  const handleFirstImageMessage = useCallback(async () => {
+    const params = new URLSearchParams(window.location.search);
+    if (
+      params.get("imageUrl") &&
+      params.get("imageName") &&
+      params.get("imageType") &&
+      params.get("imageSize")
+    ) {
+      const queryParams: { [key: string]: string } = {};
+      params.forEach((value, key) => {
+        queryParams[key] = value;
+      });
+      const ID = nanoid();
+      const message: Message = {
+        id: ID,
+        role: "user",
+        content: incomingInput || "",
+        name: `${props.username},${props.userId}`,
+      };
+      const createFileFromBlobUrl = async (
+        blobUrl: string,
+        fileName: string,
+      ) => {
+        const response = await fetch(blobUrl);
+        const blob = await response.blob();
+        return new File([blob], fileName, { type: blob.type });
+      };
+
+      const imageUrl = params.get("imageUrl")!;
+      const imageExtension = params.get("imageExtension")!;
+
+      const file = await createFileFromBlobUrl(
+        imageUrl,
+        `image.${imageExtension}`,
+      );
+      console.log("Created file from blob URL:", file);
+      const zodMessage: any = Schema.safeParse({
+        imageName: params.get("imageName"),
+        imageType: params.get("imageType"),
+        imageSize: Number(params.get("imageSize")),
+        file: file,
+        value: incomingInput || "",
+        userId: props.userId,
+        orgId: props.orgId,
+        chatId: props.chatId,
+        message: [message],
+        id: ID,
+        chattype: chattype,
+      });
+      console.log("zodMessageImage Extension:", imageExtension);
+      // console.log("zodmessage", zodMessage);
+      // console.log("dropzone", props.dropZoneActive);
+      console.log("zodMessage", zodMessage, imageExtension);
+      if (zodMessage.success) {
+        const zodMSG = JSON.stringify(zodMessage);
+        const formData = new FormData();
+        formData.append("zodMessage", zodMSG);
+        formData.append("file", file);
+        setIsRagLoading(true);
+        const response = await fetch("/api/imageInput", {
+          method: "POST",
+          body: formData,
+        });
+        if (response && response.status.toString().startsWith("2")) {
+          console.log("responce", response);
+          let assistantMsg = "";
+          const reader = response.body?.getReader();
+          console.log("reader", reader);
+          const decoder = new TextDecoder();
+          let charsReceived = 0;
+          let content = "";
+          reader
+            ?.read()
+            .then(async function processText({ done, value }) {
+              if (done) {
+                setIsRagLoading(false);
+                console.log("Stream complete");
+                return;
+              }
+              charsReceived += value.length;
+              const chunk = decoder.decode(value, { stream: true });
+              assistantMsg += chunk === "" ? `${chunk} \n` : chunk;
+              content += chunk === "" ? `${chunk} \n` : chunk;
+              // console.log("assistMsg", assistantMsg);
+              props.setMessages([
+                ...props.messages,
+                awsImageMessage,
+                message,
+                {
+                  ...assistantMessage,
+                  content: assistantMsg,
+                },
+              ]);
+              reader.read().then(processText);
+            })
+            .then((e) => {
+              setIsRagLoading(false);
+              console.error("error", e);
+            });
+          const awsImageMessage = {
+            role: "user",
+            subRole: "input-image",
+            content: `${process.env.NEXT_PUBLIC_IMAGE_PREFIX_URL}imagefolder/${props.chatId}/${ID}.${imageExtension}`,
+            id: ID,
+          } as Message;
+          const assistantMessage: Message = {
+            id: ID,
+            role: "assistant",
+            content: content,
+          };
+
+          console.log("image chat", queryParams);
+          // image chat
+        } else {
+          //TODO: api thrown some error
+          setIsRagLoading(false);
+        }
+      }
+    }
+  }, []);
+
+  useEffect(() => {
+    if (isNewChat === "true" && incomingInput) {
+      //TODO: use types for useQueryState
+      if (incomingInput && chattype !== "tldraw") {
+        const params = new URLSearchParams(window.location.search);
+        if (
+          params.get("imageUrl") &&
+          params.get("imageName") &&
+          params.get("imageType") &&
+          params.get("imageSize")
+        ) {
+          console.log("zodMessage", "we made to here", params);
+          handleFirstImageMessage();
+        } else {
+          const newMessage = {
+            id: nanoid(),
+            role: "user",
+            content: incomingInput,
+            name: `${props.username},${props.userId}`,
+            audio: "",
+          } as Message;
+          props.append(newMessage);
+        }
+      }
+    }
+    setIsFromClipboard("false");
+    setIsNewChat("false");
+  }, [isFromClipboard, isNewChat]);
   // const ably = useAbly();
 
   // console.log(
@@ -319,30 +477,6 @@ const InputBar = (props: InputBarProps) => {
     props.setInput("");
   };
 
-  const handleAudio = async (audioFile: File) => {
-    setIsAudioWaveVisible(false);
-    setIsTranscribing(true);
-    const f = new FormData();
-    f.append("file", audioFile);
-    // Buffer.from(audioFile)
-    console.log(audioFile);
-    try {
-      const res = await fetch("/api/transcript", {
-        method: "POST",
-        body: f,
-      });
-
-      // console.log('data', await data.json());
-      const data = await res.json();
-      console.log("got the data", data);
-      props.setInput(data.text);
-      setIsTranscribing(false);
-    } catch (err) {
-      console.error("got in error", err);
-      setIsTranscribing(false);
-    }
-  };
-
   const [audioId, setAudioId] = useState(0);
   const [transcriptHashTable, setTranscriptHashTable] = useState<{
     [key: number]: string;
@@ -417,7 +551,9 @@ const InputBar = (props: InputBarProps) => {
         if (countdown > 0) {
           updateStatus({
             isTyping: true,
-            username: `Echoes is thinking (${countdown--} secs)`,
+            username: props?.isLoading
+              ? `Echoes is thinking (${countdown--} secs)`
+              : `Echoes is typing (${countdown--} secs)`,
             id: props.userId,
           });
         } else {
